Рефакторинг компонента CustomDivider.
